COMMUNICATION-1187 General-Block-Transfer TypedProperty returns BigDecimal value
git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@90254 0d108963-8a47-0410-809b-ebd9fb3b3407

+++ upl/com.energyict.mdc.upl.impl/protocolimpl-v2/src/main/java/com/energyict/dlms/common/DlmsProtocolProperties.java
@@ -100,7 +100,7 @@ public abstract class DlmsProtocolProperties implements DlmsSessionProperties {
     public static final boolean DEFAULT_FIX_MBUS_HEX_SHORT_ID = false;
     public static final String DEFAULT_DEVICE_ID = "";
     public static final boolean DEFAULT_ENABLE_GBT = false;
-    public static final int DEFAULT_GBT_WINDOW_SIZE = -1;
+    public static final BigDecimal DEFAULT_GBT_WINDOW_SIZE = BigDecimal.valueOf(-1);
 
     private TypedProperties typedProperties;
